Question Answering Application/requirements.txt
langchain
openai
streamlit
langchain-openai
python-dotenv
ipykernel
transformers
torch
sentencepiece